requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9
itsdangerous>=2.2.0
starlette>=0.40.0
anthropic>=0.76.0
python-dotenv>=1.0.1
requests>=2.31.0
stripe>=8.0.0
dnspython>=2.6.0
alembic>=1.13.2
celery[redis]>=5.4.0
redis>=5.0.4
sentry-sdk>=2.3.1
structlog>=24.1.0
